Include stored pages in rowgroup size
Store the size of the page in the data page struct when it's added to the column store. And use that value when calculating the size of a Column. This ensures we keep track of the total size of the current row group to know when to flush it to disk.